refactor: fix folder paths
