+++ src/obsidian/dashboards-settings-tab.ts/styles.css
@@ -1,0 +1,3 @@
+.Dashboards__setting-emphasize {
+    color: var(--text-accent);
+}
