+++ src/obsidian/layout-modal/index.ts
@@ -2,7 +2,7 @@ import { App, Modal } from "obsidian";
 import { deserializeAppState } from "src/data/serialize";
 import { getGridX, getGridY } from "src/react/table/table-utils";
 import { AppState, Container } from "src/shared/state/types";
-import { updateContainerPosition } from "../modal/array-utils";
+import { updateContainerPosition } from "../array-utils";
 
 import "./styles.css";
