+++ src/react/container/empty-container-content.tsx
@@ -1,5 +1,5 @@
 import LinkModal from "src/obsidian/link-modal";
-import FileModal from "src/obsidian/modal/file-modal";
+import FileModal from "src/obsidian/file-modal";
 import IconButton from "../icon-button/icon-button";
 import { TFile } from "obsidian";
 import CodeBlockModal from "src/obsidian/code-block-modal";
